New tries
tried to attach body movements to the mosquito thing, but didnt succeed. tried to make the mosquitos find the color of black that is the color of the contour of the body but seems like a bad idea so now im trying a different approach where i use the kinect examples of shiffman but hasnt found out much the examples by shiffman also use a different library.
mygadvanced is to isolate the mosquitos code and make it more advanced hasnt done much on it yet though...